Import symbols rather than qualifying them
Many symbols from other packages were qualified.
For instance "LinearAlgebra.tr". This commit removes
the qualifications and imports the symbols explicitly
at the top of the files.

"export Package: func" for functions that will be used,
but not extended.

"import Package.func" for symbols that will be extended.
The only instance of this was present before this commit.

"import Package: func" (or Package.func) for symbols that
will be used, but are not exported by Package.

"using QuantumInfo" was present before the upgrade to v1.0,
so I left it intact.